+++ packages/stream_video_flutter/lib/src/theme/components/components.dart
@@ -1,5 +1,6 @@
 export 'call_app_bar_theme.dart';
 export 'call_control_bar_theme.dart';
+export 'call_duration_badge_theme.dart';
 export 'call_participants_grid_theme.dart';
 export 'connection_quality_indicator_theme.dart';
 export 'floating_participant_tile_theme.dart';